% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{classification_results}
\alias{classification_results}
\title{Example of output from designSampleSizeClassification function}
\format{
A list with six elements
}
\usage{
classification_results
}
\description{
It is the output of \code{\link{designSampleSizeClassification}} function
with a list of \code{\link{simulated_datasets}}
generated under same protein number and sample size.
The list should include the required elements as below.
}
\details{
\itemize{
  \item num_proteins : the number of simulated proteins
  \item num_samples : a vector with the number of simulated samples in each condition
  \item mean_predictive_accuracy : the mean predictive accuracy over all the simulated datasets.
  \item mean_feature_importance : the mean protein importance vector over all the simulated datasets,
  the length of which is `num_proteins'.
  \item predictive_accuracy : a vector of predictive accuracy on each simulated dataset.
  \item feature_importance : a matrix of feature importance,
  where rows are proteins and columns are simulated datasets.
  the length of which is `num_proteins'.
}
}
\examples{
classification_results$num_proteins
classification_results$num_samples
classification_results$mean_predictive_accuracy
head(classification_results$mean_feature_importance)

}
\keyword{datasets}
