% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{.catch_faults}
\alias{.catch_faults}
\title{Catch Faults}
\usage{
.catch_faults(..., conn, session = NULL)
}
\arguments{
\item{conn}{A connection object generated by .logGeneration}

\item{session}{A session object for using with the shiny server}
}
\description{
A wrapper function that catches exceptions and logs them appropriately
}
\keyword{internal}
