% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{.classification_model}
\alias{.classification_model}
\title{Fit a classification model}
\usage{
.classification_model(x, y, classifier, tunegrid, ...)
}
\arguments{
\item{x}{protein abundance data for one protein}

\item{y}{group information}

\item{classifier}{A string of classifier}

\item{tunegrid}{A data.frame with tuning parameters}
}
\value{
trained classification model
}
\description{
Fit a classification model
}
\keyword{internal}
