% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{.classification_performance}
\alias{.classification_performance}
\title{For each simulated dataset, calculate predictive accuracy on validation set}
\usage{
.classification_performance(
  index,
  classifier,
  train_x_list,
  train_y_list,
  valid_x,
  valid_y,
  top_K,
  tunegrid
)
}
\arguments{
\item{index}{protein abundance data for one protein.}

\item{classifier}{A character string with the name of the classifier to be used}
}
\value{
A list with (1) predictive accuracy on validation set and
(2) the important features
}
\description{
For each simulated dataset, calculate predictive accuracy on validation set
}
\keyword{internal}
