% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{.classificationModel}
\alias{.classificationModel}
\title{Fit a classification model}
\usage{
.classificationModel(x, y, classifier)
}
\arguments{
\item{x}{protein abundance data for one protein}

\item{y}{group information}
}
\value{
trained classification model
}
\description{
Fit a classification model
}
\keyword{internal}
