% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{.classificationPerformance}
\alias{.classificationPerformance}
\title{For each simulated dataset, calculate predictive accuracy on validation set}
\usage{
.classificationPerformance(
  index,
  classifier,
  train_x_list,
  train_y_list,
  valid_x,
  valid_y,
  top_K
)
}
\arguments{
\item{index}{protein abundance data for one protein.}
}
\value{
A list with (1) predictive accuracy on validation set and
(2) the trained classification model
}
\description{
For each simulated dataset, calculate predictive accuracy on validation set
}
\keyword{internal}
