% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{.data_checks}
\alias{.data_checks}
\title{Data Checking}
\usage{
.data_checks(data, annotation, conn)
}
\arguments{
\item{data}{A abundance file input}

\item{annotation}{Annotation file for the abundance}

\item{conn}{A connection}
}
\description{
This function check for data consistency in the input dataset
and provided either a failure or success message when the data checks are
passed successfully
}
\keyword{internal}
