% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{.do_prcomp}
\alias{.do_prcomp}
\title{Do Principal Component Analysis}
\usage{
.do_prcomp(sim_x, sim_y)
}
\arguments{
\item{sim_x}{A dataframe of the feature variables}

\item{sim_y}{A dataframe of the predictor vairable}
}
\value{
A named list of the outputs
}
\description{
A wrapper function to the base `prcomp()` formats the results
in a required output format
}
\keyword{internal}
