% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{.format_summary_table}
\alias{.format_summary_table}
\title{Get summary table for the annotation data}
\usage{
.format_summary_table(data = NULL)
}
\arguments{
\item{data}{A data.frame with the annotation data containing the Bioreplicates
Runs and condition information}
}
\value{
A data.frame with the counts of bioreplicates for each conditions
and the number of runs as well
}
\description{
Get the summary for the unique bioreplicates and number of
MS runs for the data that is provided
}
