% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{.identify_optimal}
\alias{.identify_optimal}
\title{Identify optimal sample size}
\usage{
.identify_optimal(df, cutoff, use_h2o = F)
}
\arguments{
\item{df}{A dataframe containing the sample size, accuracies}

\item{cutoff}{A numeric values to determine the threshold for identifying the
sample size}

\item{use_h2o}{A logical input to use logic for h2o data}
}
\value{
\emph{df} A data.frame

\emph{opt} A integer with the optimal sample size

\emph{y_lim} A vector of limits for a ggplot2 object
}
\description{
Identify optimal sample size
}
\keyword{internal}
