% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{.log_write}
\alias{.log_write}
\title{Write Log}
\usage{
.log_write(log, log_type, conn)
}
\arguments{
\item{log}{A string log that has to be written to the log file}

\item{log_type}{A type of log to be written (purely for informational purpose)}

\item{conn}{A connection object to the logfile}
}
\description{
Write Log
}
\keyword{internal}
