% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{.logGeneration}
\alias{.logGeneration}
\title{Create log file}
\usage{
.logGeneration(...)
}
\value{
\emph{finalfile} The log file name

\emph{processout} The current log information
}
\description{
Create log file
}
\keyword{internal}
