% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{.pca_plot}
\alias{.pca_plot}
\title{Plot PCA outputs}
\usage{
.pca_plot(data, exp_var, ...)
}
\arguments{
\item{data}{A data frame containing the Principal components to be plotted}

\item{exp_var}{A vector containing numeric values of the expected variance}
}
\description{
A utility wrapper for plotting pca data as returned by the `do_prcomp()`
}
