% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{.quantile_eval}
\alias{.quantile_eval}
\title{Evaluate quantiles given vectors and probability}
\usage{
.quantile_eval(df, vec_name, prob, eq, rnd = 4)
}
\arguments{
\item{df}{A data.frame, with column names for which quantiles are calculated}

\item{vec_name}{A vector with the name of the column}

\item{prob}{A numeric value for the probablitiy}

\item{rnd}{A integer value which defines where to round the quantiles}
}
\description{
Evaluate quantiles given vectors and probability
}
