% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{.randomImputation}
\alias{.randomImputation}
\title{For each protein, impute the missing values based on the observed values}
\usage{
.randomImputation(data)
}
\arguments{
\item{data}{protein abundance data for one protein.}
}
\value{
Imputed protein abundance data
}
\description{
For each protein, impute the missing values based on the observed values
}
\keyword{internal}
