% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{.sampleSimulation}
\alias{.sampleSimulation}
\title{Simulate extended datasets for sample size estimation}
\usage{
.sampleSimulation(m, mu, sigma)
}
\arguments{
\item{m}{number of samples per group to simulate}

\item{mu}{a matrix of mean abundance in each phenotype group and each protein}

\item{sigma}{a matrix of variance in each phenotype group and each protein}
}
\value{
\emph{X} A simulated matrix with required sample size

\emph{Y} Group information corresponding with \emph{X}
}
\description{
Simulate extended datasets for sample size estimation
}
\keyword{internal}
