% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{.status}
\alias{.status}
\title{Status updates}
\usage{
.status(detail, ...)
}
\arguments{
\item{detail}{Message that need to be logged/displayed}
}
\description{
Logs all messages comming through to the log file and to the console
also provided status updates to the shiny interface with the right options
}
\keyword{internal}
