% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateVar.R
\name{estimateVar}
\alias{estimateVar}
\title{Estimate the mean abundance and variance of each protein in each condition.}
\usage{
estimateVar(data, annotation)
}
\arguments{
\item{data}{Data matrix with protein abundance.
Rows are proteins and columns are Biological replicates or samples.}

\item{annotation}{Group information for samples in data.
`Run' for MS run, `BioReplicate' for biological subject ID and `Condition' for group information are required.
`Run' information should be the same with the column of `data'. Multiple `Run' may come from same `BioReplicate'.}
}
\value{
\emph{model} is the list of linear models trained for each protein.

\emph{mu} is the mean abundance matrix of each protein in each phenotype group.

\emph{sigma} is the sd matrix of each protein in each phenotype group.

\emph{promean} is the mean abundance vector of each protein across all the samples.

\emph{protein} is proteins, correpsonding to the rows in \emph{mu} and \emph{sigma} or the element of \emph{promean}.
}
\description{
Estimate the mean abundance and variance of each protein in each condition.
}
\details{
The function fits intensity-based linear model on the input data `data'.
This function outputs variance components and mean abundance for each protein.
}
\examples{
data(OV_SRM_train)
data(OV_SRM_train_annotation)

# estimate the mean protein abunadnce and variance in each condition
variance_estimation <- estimateVar(data = OV_SRM_train,
                                   annotation = OV_SRM_train_annotation)

# the mean protein abundance in each condition
head(variance_estimation$mu)

# the standard deviation in each condition
head(variance_estimation$sigma)

# the mean protein abundance across all the conditions
head(variance_estimation$promean)

}
\author{
Ting Huang, Meena Choi, Olga Vitek
}
