% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_methods.R
\name{h2o_config}
\alias{h2o_config}
\title{H2o configuration}
\usage{
h2o_config(...)
}
\value{
A named list containing required h2o configuration details, if none 
provided, defaults are used
}
\description{
Get configuration details of the h2o instance to start, the details
can be specified in the Rprofile.site file in the R installation path
}
\keyword{internal}
