% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanSDplot.R
\name{meanSDplot}
\alias{meanSDplot}
\title{Mean-SD plot}
\usage{
meanSDplot(
  data,
  smoother_size = 1,
  xlimUp = 30,
  ylimUp = 3,
  save.pdf = FALSE,
  ...
)
}
\arguments{
\item{data}{A list with mean protein abundance vector and standard deviation vector.
It should be the output of \code{\link{estimateVar}} function.}

\item{smoother_size}{Size of lowess smoother. Default is 1.}

\item{xlimUp}{The upper limit of x-axis for mean-SD plot. Default is 30.}

\item{ylimUp}{The upper limit of y-axis for mean-SD plot. Default is 3.}

\item{save.pdf}{A logical input, determines to save the plots as a pdf or not,
the pdf plot is saved in the current working directory, name of the created
file is displayed on the console and logged for easier access}
}
\value{
\emph{meanSDplot} is the plot for the mean protein abundance (X-axis)
vs standard deviation (Y-axis).
}
\description{
Draw the plot for the mean protein abundance vs standard deviation
(square root of variance). The `lowess` function is used to fit the LOWESS 
smoother between mean protein abundance and standard deviation.
}
\examples{
data(OV_SRM_train)
data(OV_SRM_train_annotation)

variance_estimation <- estimateVar(data = OV_SRM_train,
                                   annotation = OV_SRM_train_annotation)

meanSDplot(variance_estimation)

}
\author{
Ting Huang, Meena Choi, Olga Vitek
}
