% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{OV_SRM_train}
\alias{OV_SRM_train}
\title{The training set from a study for subjects with ovarian cancer}
\format{A numeric matrix with 68 rows and 260 columns.}
\usage{
OV_SRM_train
}
\description{
It is a protein abundance data matrix,
where rows are proteins and columns are samples.
It includes log2 protein intensities for 68 proteins
among 260 biological subjects
It is the input for estimateVar and simulateDataset function,
with annotation file. It should be prepared by users.
}
\examples{
head(OV_SRM_train)

}
\references{
Huttenhain R and Choi M et al. (2019).
 A targeted mass spectrometry strategy for developing proteomic biomarkers:
 a case study of epithelial ovarian cancer.
Molecular & Cellular Proteomics 18(9):1836-1850. doi:10.1074/mcp.RA118.001221.
}
\keyword{datasets}
