% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{OV_SRM_train_annotation}
\alias{OV_SRM_train_annotation}
\title{Annotation file for \code{\link{OV_SRM_train}},}
\format{A data frame with 173 rows and 2 variables.}
\usage{
OV_SRM_train_annotation
}
\description{
Annotation of example data, \code{\link{OV_SRM_train}}, in this package.
It should be prepared by users. The variables are as follows:
}
\details{
\itemize{
  \item BioReplicate : Unique ID for biological subject.
  It should be the same as the column names of \code{\link{OV_SRM_train}}
  \item Condition : Condition for BioReplicate (ex. Healthy, Cancer, Time0)
}
}
\examples{
head(OV_SRM_train_annotation)

}
\references{
Huttenhain R and Choi M et al. (2019).
 A targeted mass spectrometry strategy for developing proteomic biomarkers:
 a case study of epithelial ovarian cancer.
 Mol Cell Proteomics 18(9):1836-1850.
 doi:10.1074/mcp.RA118.001221.
}
\keyword{datasets}
