% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{qc_boxplot}
\alias{qc_boxplot}
\title{Plot QC boxplots}
\usage{
qc_boxplot(data = NULL, annot = NULL)
}
\arguments{
\item{data}{A formatted data.frame with abundance, bioreplicate and conditions 
information}
}
\value{
A plotly object
}
\description{
Plots an interactive plotly graphic with boxplots for all the 
proteins and their respective conditions with abundance indicated in the log
scale on the y axis
}
\keyword{internal}
