% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{simulate_grid}
\alias{simulate_grid}
\title{Simulate datasets to be tested out}
\usage{
simulate_grid(
  data = NULL,
  annot = NULL,
  num_simulation,
  exp_fc,
  list_diff_proteins,
  samples_per_group,
  sim_valid,
  rank_proteins,
  mean_ip,
  mean_equality,
  sd_ip,
  sd_equality,
  valid_samples_per_grp,
  seed,
  est_var,
  conn,
  session = NULL
)
}
\description{
A wrapper function for the `simulateDataset` function from the 
MSstatsSampleSize package which enables simulating datasets for running experiments
}
