% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{ss_classify_caret}
\alias{ss_classify_caret}
\title{Classsification caret}
\usage{
ss_classify_caret(n_samp, sim_data, classifier, k = 10, session = NULL, ...)
}
\arguments{
\item{n_samp}{A character vector consisting the of the different samples simulated}

\item{sim_data}{A list object, output out the `simulate_grid` function}

\item{classifier}{A string specifying the classifier to use}

\item{k}{A integer value for the number of features to select}

\item{session}{A session object for the shiny app}
}
\description{
A wrapper function to classify multiple simulated datasets
}
\keyword{internal}
