% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_methods.R
\name{ss_classify_h2o}
\alias{ss_classify_h2o}
\title{Classify using h2o classification algorithms}
\usage{
ss_classify_h2o(
  n_samp,
  sim_data,
  classifier,
  stopping_metric = "AUTO",
  seed = -1,
  nfolds = 0,
  fold_assignment = "AUTO",
  iters = 200,
  alpha = 0,
  family,
  solver,
  link,
  min_sdev,
  laplace,
  eps,
  session = NULL
)
}
\arguments{
\item{n_samp}{A comma separated chracter vector with number of sample sizes}

\item{sim_data}{A list object containing simulated data}

\item{classifier}{Classifier to use as selected by the user can use one of the
following `Logistic Regression`, `Random Forest`, `Support Vector Machine`,
`Naive Bayes`, `Neural Network`}

\item{stopping_metric}{}

\item{session}{}
}
\description{
This function enables the shiny application to apply various 
algorithms as defined by the user
}
