% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{theme_MSstats}
\alias{theme_MSstats}
\title{MSstats Theme}
\usage{
theme_MSstats(
  x.axis.size = 10,
  y.axis.size = 10,
  legend.size = 11,
  margin = 0.5,
  leg.dir = "horizontal",
  download = F,
  ...
)
}
\arguments{
\item{x.axis.size}{A numeric value for size for the elements on the x-axis}

\item{y.axis.size}{A numeric value for size for the elements on the y-axis}

\item{legend.size}{A numeric value fot the size of the legend}
}
\description{
A utility function that standardized all the ggplot themes
}
\keyword{internal}
