% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{variance_estimation}
\alias{variance_estimation}
\title{Example of output from estimateVar function}
\format{A list with four elements}
\usage{
variance_estimation
}
\description{
It is the output of \code{\link{estimateVar}} function with two inputs:
\code{\link{OV_SRM_train}} and \code{\link{OV_SRM_train_annotation}}.
The list should include the required elements as below.
}
\details{
\itemize{
  \item mu : the mean abundance matrix of each protein in each condition
  \item sigma : the standard deviation matrix of each protein in each condition
  \item promean: the mean abundance vector of each protein across all the samples.
  \item protein : proteins, correpsonding to the rows in \emph{mu} and \emph{sigma} or the element of \emph{promean}
}
}
\examples{
head(variance_estimation$mu)
head(variance_estimation$sigma)
head(variance_estimation$promean)

}
\keyword{datasets}
