% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{.logGeneration}
\alias{.logGeneration}
\title{Create log file}
\usage{
.logGeneration(...)
}
\value{
\emph{con} A file connection to write the logs into

\emph{file} Name of the file to which connection is open
}
\description{
Create log file
}
\keyword{internal}
